+++ utils/model.py
@@ -1,0 +1,6 @@
+users:list=[
+    {'name':'maciej','location':'łódź','posts':100},
+    {'name':'mateusz','location':'łódź','posts':200},
+    {'name':'maciej81','location':'łódź','posts':300},
+    {'name':'kondrad','location':'łódź','posts':400},
+]
